Fill content of Game 3. Using \def instead of copy paste.
Content of Game 3 is now complete. I use the \def command to insert the
textareas from the categoies/file.tex file instead of copy&paste them
into the main file. To do that, the files have to be included via the
\input{file.tex} command and each textarea in the file has to be given a
unique variable name with the \def \variable {textarea...} command. The
variable names are then called in the content of the main file. The
variable names are ordered by \XY where X and Y are in {a,...,e} and X
corresponds to the subject (1 to 5) while Y corresponds to the Price
(100 to 500). This is necessary, because the \def command does not allow
for numbers inside variable names.